chagne language get data from server mixin

+++ lib/data/models/product_model/product_model.dart
@@ -12,7 +12,7 @@ class ProductModel {
   String? brandName;
   int? inStock;
   bool? isActive;
-  int? masterPrice;
+  double? masterPrice;
   String? locale;
 
   String? country;
